I added the Homophone cipher code to the MainFrame class
